+++ mlflow/gateway/schemas/embeddings.py
@@ -1,7 +1,7 @@
 from typing import Optional, Union
 
 from mlflow.gateway.base_models import RequestModel, ResponseModel
-from mlflow.utils import IS_PYDANTIC_V2
+from mlflow.utils import IS_PYDANTIC_V2_OR_NEWER
 
 _REQUEST_PAYLOAD_EXTRA_SCHEMA = {
     "example": {
@@ -14,7 +14,7 @@ class RequestPayload(RequestModel):
     input: Union[str, list[str], list[int], list[list[int]]]
 
     class Config:
-        if IS_PYDANTIC_V2:
+        if IS_PYDANTIC_V2_OR_NEWER:
             json_schema_extra = _REQUEST_PAYLOAD_EXTRA_SCHEMA
         else:
             schema_extra = _REQUEST_PAYLOAD_EXTRA_SCHEMA
@@ -91,7 +91,7 @@ class ResponsePayload(ResponseModel):
     usage: EmbeddingsUsage
 
     class Config:
-        if IS_PYDANTIC_V2:
+        if IS_PYDANTIC_V2_OR_NEWER:
             json_schema_extra = _RESPONSE_PAYLOAD_EXTRA_SCHEMA
         else:
             schema_extra = _RESPONSE_PAYLOAD_EXTRA_SCHEMA
