+++ mlflow/langchain/utils/chat.py
@@ -13,7 +13,7 @@ from mlflow.environment_variables import (
 )
 from mlflow.exceptions import MlflowException
 from mlflow.types.schema import Array, ColSpec, DataType, Schema
-from mlflow.utils import IS_PYDANTIC_V2
+from mlflow.utils import IS_PYDANTIC_V2_OR_NEWER
 
 _logger = logging.getLogger(__name__)
 
@@ -154,7 +154,7 @@ def try_transform_response_to_chat_format(response):
         ),
     )
 
-    if IS_PYDANTIC_V2:
+    if IS_PYDANTIC_V2_OR_NEWER:
         return transformed_response.model_dump(mode="json")
     else:
         return json.loads(transformed_response.json())
@@ -180,7 +180,7 @@ def try_transform_response_iter_to_chat_format(chunk_iter):
             ],
         )
 
-        if IS_PYDANTIC_V2:
+        if IS_PYDANTIC_V2_OR_NEWER:
             return transformed_response.model_dump(mode="json")
         else:
             return json.loads(transformed_response.json())
@@ -216,7 +216,7 @@ def try_transform_response_iter_to_chat_format(chunk_iter):
 
 
 def _convert_chat_request_or_throw(chat_request: dict):
-    if IS_PYDANTIC_V2:
+    if IS_PYDANTIC_V2_OR_NEWER:
         model = _ChatRequest.model_validate(chat_request)
     else:
         model = _ChatRequest.parse_obj(chat_request)
