+++ tests/tracing/test_fluent.py
@@ -1475,55 +1475,3 @@ def test_add_trace_logging_model_from_code():
     trace = mlflow.get_last_active_trace()
     assert trace is not None
     assert len(trace.data.spans) == 2
-
-
-def test_set_span_chat_messages_and_tools():
-    messages = [
-        {
-            "role": "system",
-            "content": "please use the provided tool to answer the user's questions",
-        },
-        {"role": "user", "content": "what is 1 + 1?"},
-        {
-            "role": "assistant",
-            "tool_calls": [
-                {
-                    "id": "123",
-                    "function": {"arguments": '{"a": 1,"b": 2}', "name": "add"},
-                    "type": "function",
-                }
-            ],
-        },
-    ]
-
-    tools = [
-        {
-            "type": "function",
-            "function": {
-                "name": "add",
-                "description": "Add two numbers",
-                "parameters": {
-                    "type": "object",
-                    "properties": {
-                        "a": {"type": "number"},
-                        "b": {"type": "number"},
-                    },
-                    "required": ["a", "b"],
-                },
-            },
-        }
-    ]
-
-    @mlflow.trace(span_type=SpanType.CHAT_MODEL)
-    def dummy_call(messages, tools):
-        span = mlflow.get_current_active_span()
-        mlflow.set_span_chat_messages(span, messages)
-        mlflow.set_span_chat_tools(span, tools)
-        return None
-
-    dummy_call(messages, tools)
-
-    trace = mlflow.get_last_active_trace()
-    span = trace.data.spans[0]
-    assert span.get_attribute(SpanAttributeKey.CHAT_MESSAGES) == messages
-    assert span.get_attribute(SpanAttributeKey.CHAT_TOOLS) == tools
